Fix addNewAction function in EditedRepository - Fixed null check: changed 'insertionIndex != null' to 'insertionIndex == null'   (when index is null, action should be added at end of list) - Fixed priority calculation: changed 'previousActions.lastIndex * 1' to 'previousActions.size'   (size gives correct next priority index) - Fixed method call: changed 'updateScenario()' to 'updatePriorities()'   (correct method for updating priorities of subsequent actions)

+++ app/src/main/java/com/example/clicka/config/data/EditedPreferences.kt
@@ -28,8 +28,8 @@ internal fun SharedPreferences.Editor.putClickRepeatCountConfig(durationMs: Long
     putLong(PREF_LAST_CLICK_REPEAT_COUNT,durationMs)
 
 /** return the default repeat delay for a click */
-internal fun SharedPreferences.getClickRepeatDelayConfig(default: Int):Int =
-    getInt(PREF_LAST_CLICK_REPEAT_DELAY, default)
+internal fun SharedPreferences.getClickRepeatDelayConfig(default: Long):Long =
+    getLong(PREF_LAST_CLICK_REPEAT_DELAY, default)
 
 /** Save a new default repeat delay for a click*/
 internal fun SharedPreferences.Editor.putClickRepeatDelayConfig(durationMs: Long): SharedPreferences.Editor=
